refactor(utils): extract filesystem helpers into mtui/fileops.py
Move chdir re-export, ensure_dir_exists, atomic_write_file, and
timestamp verbatim out of mtui/utils.py into a new mtui/fileops.py
module. utils.py keeps a re-export shim for callers so this commit is
byte-identical at the public-API surface.

Third commit of the Phase 5b / C8 split (PLAN.md item 11).
